chore(history): rename history_controller.rs to controller.rs
Aligns history with connect (which already uses controller.rs) ahead of
splitting it into a controller/ submodule folder.
