+++ rust/operator-binary/src/history/controller.rs
